projects/requirements.txt
pandas
numpy
scikit-learn
boto3
mlflow
imblearn
xgboost
pingouin
python-dotenv
